+++ src/endpoint.ts
@@ -26,7 +26,7 @@ import { AbstractLogger } from "./logger";
 import { LogicalContainer, combineContainers } from "./logical-container";
 import { AuxMethod, Method } from "./method";
 import { AnyMiddlewareDef } from "./middleware";
-import { mimeJson, mimeMultipart, mimeRaw } from "./mime";
+import { ContentType, contentTypes } from "./content-type";
 import { AnyResultHandlerDefinition } from "./result-handler";
 import { Security } from "./security";
 
@@ -63,6 +63,7 @@ export abstract class AbstractEndpoint {
   public abstract getScopes(): ReadonlyArray<string>;
   public abstract getTags(): ReadonlyArray<string>;
   public abstract getOperationId(method: Method): string | undefined;
+  public abstract getRequestType(): ContentType;
 }
 
 export class Endpoint<
@@ -86,6 +87,7 @@ export class Endpoint<
   readonly #scopes: ReadonlyArray<SCO>;
   readonly #tags: ReadonlyArray<TAG>;
   readonly #getOperationId: (method: Method) => string | undefined;
+  readonly #requestType: ContentType;
 
   constructor({
     methods,
@@ -133,13 +135,13 @@ export class Endpoint<
     this.#responses = {
       positive: Object.freeze(
         normalizeApiResponse(resultHandler.getPositiveResponse(outputSchema), {
-          mimeTypes: [mimeJson],
+          mimeTypes: [contentTypes.json],
           statusCodes: [defaultStatusCodes.positive],
         }),
       ),
       negative: Object.freeze(
         normalizeApiResponse(resultHandler.getNegativeResponse(), {
-          mimeTypes: [mimeJson],
+          mimeTypes: [contentTypes.json],
           statusCodes: [defaultStatusCodes.negative],
         }),
       ),
@@ -152,14 +154,13 @@ export class Endpoint<
         ),
       );
     }
+    this.#requestType = hasUpload(inputSchema)
+      ? "upload"
+      : hasRaw(inputSchema)
+        ? "raw"
+        : "json";
     this.#mimeTypes = {
-      input: Object.freeze(
-        hasUpload(inputSchema)
-          ? [mimeMultipart]
-          : hasRaw(inputSchema)
-            ? [mimeRaw]
-            : [mimeJson],
-      ),
+      input: Object.freeze([contentTypes[this.#requestType]]),
       positive: Object.freeze(
         this.#responses.positive.flatMap(({ mimeTypes }) => mimeTypes),
       ),
@@ -191,6 +192,10 @@ export class Endpoint<
 
   public override getMimeTypes(variant: MimeVariant) {
     return this.#mimeTypes[variant];
+  }
+
+  public override getRequestType() {
+    return this.#requestType;
   }
 
   public override getResponses(variant: ResponseVariant) {