+++ src/file-schema.ts
@@ -1,17 +1,20 @@
 import { z } from "zod";
-import { proprietary } from "./metadata";
 
-export const ezFileKind = "File";
+export const ezFileBrand = Symbol("File");
 
 const bufferSchema = z.custom<Buffer>((subject) => Buffer.isBuffer(subject), {
   message: "Expected Buffer",
 });
 
 const variants = {
-  buffer: () => proprietary(ezFileKind, bufferSchema),
-  string: () => proprietary(ezFileKind, z.string()),
-  binary: () => proprietary(ezFileKind, bufferSchema.or(z.string())),
-  base64: () => proprietary(ezFileKind, z.string().base64()),
+  buffer: () => bufferSchema.brand(ezFileBrand as symbol),
+  string: () => z.string().brand(ezFileBrand as symbol),
+  binary: () => bufferSchema.or(z.string()).brand(ezFileBrand as symbol),
+  base64: () =>
+    z
+      .string()
+      .base64()
+      .brand(ezFileBrand as symbol),
 };
 
 type Variants = typeof variants;
@@ -22,3 +25,5 @@ export function file<K extends Variant>(variant: K): ReturnType<Variants[K]>;
 export function file<K extends Variant>(variant?: K) {
   return variants[variant || "string"]();
 }
+
+export type FileSchema = ReturnType<typeof file>;