+++ tests/unit/checks.spec.ts
@@ -3,13 +3,13 @@ import { describe, expect, test, vi } from "vitest";
 import { z } from "zod";
 import { ez } from "../../src";
 import { hasNestedSchema, hasTransformationOnTop } from "../../src/deep-checks";
-import { isProprietary } from "../../src/metadata";
-import { ezUploadKind } from "../../src/upload-schema";
+import { metaSymbol } from "../../src/metadata";
+import { ezUploadBrand } from "../../src/upload-schema";
 
 describe("Checks", () => {
   describe("hasNestedSchema()", () => {
     const condition = (subject: z.ZodTypeAny) =>
-      isProprietary(subject, ezUploadKind);
+      subject._def[metaSymbol]?.brand === ezUploadBrand;
 
     test("should return true for given argument satisfying condition", () => {
       expect(hasNestedSchema({ subject: ez.upload(), condition })).toBeTruthy();